+++ netbox_healthcheck_plugin/views.py
@@ -1,7 +1,12 @@
-from health_check.views import MainView
+from health_check.views import HealthCheckView
 
 
-class HealthCheckListView(MainView):
+class HealthCheckListView(HealthCheckView):
     """Health check view that integrates django-health-check with NetBox UI."""
 
     template_name = 'netbox_healthcheck_plugin/healthcheck.html'
+    checks = [
+        'health_check.Database',
+        'netbox_healthcheck_plugin.backends.redis.NetBoxRedisCacheHealthCheck',
+        'netbox_healthcheck_plugin.backends.redis.NetBoxRedisTasksHealthCheck',
+    ]